+++ models/list.js
@@ -1,0 +1,14 @@
+var lists = [
+  {
+    "name": "WDI",
+    "todos": [1]
+  },
+  {
+    "name": "Self",
+    "todos": [2, 3, 5]
+  },
+  {
+    "name": "Errands",
+    "todos": [4]
+  }
+]
