+++ models/task.js
@@ -1,0 +1,27 @@
+var tasks = [
+  {
+    "id": 1,
+    "body": "Finish Project",
+    "completed": false
+  },
+  {
+    "id": 2,
+    "body": "Clean",
+    "completed": false
+  },
+  {
+    "id": 3,
+    "body": "Read",
+    "completed": false
+  },
+  {
+    "id": 4,
+    "body": "Groceries",
+    "completed": false
+  },
+  {
+    "id": 5,
+    "body": "Fix sink",
+    "completed": false
+  }
+]
